+++ extensions/script-kit/src/run-script.tsx
@@ -1,4 +1,14 @@
-import { ActionPanel, closeMainWindow, Icon, List, popToRoot, showToast, ToastStyle } from "@raycast/api";
+import {
+  Action,
+  ActionPanel,
+  Icon,
+  LaunchProps,
+  List,
+  Toast,
+  closeMainWindow,
+  popToRoot,
+  showToast,
+} from "@raycast/api";
 import { execa } from "execa";
 import { readFile } from "fs/promises";
 import { homedir } from "os";
@@ -7,8 +17,9 @@ import { useEffect, useState } from "react";
 import { Script } from "./script";
 
 const kit_path = resolve(homedir(), ".kit");
+const kar_path = resolve(kit_path, "kar");
 
-export default function listScripts(): JSX.Element {
+export default function listScripts(props: LaunchProps) {
   const [scripts, setScripts] = useState<Script[]>();
   useEffect(() => {
     const script_db_path = resolve(kit_path, "db", "scripts.json");
@@ -17,6 +28,13 @@ export default function listScripts(): JSX.Element {
       .then(JSON.parse)
       .then((json) => setScripts(json.scripts));
   }, []);
+
+  if (props.launchContext?.scriptPath) {
+    popToRoot();
+    closeMainWindow();
+    execa("/bin/bash", [kar_path, props.launchContext.scriptPath]);
+    return null;
+  }
 
   return (
     <List isLoading={typeof scripts == "undefined"}>
@@ -27,11 +45,21 @@ export default function listScripts(): JSX.Element {
             title={script.name}
             key={script.command}
             subtitle={script.description}
-            accessoryTitle={script.kenv}
+            accessories={[{ text: script.kenv }]}
             actions={
               <ActionPanel>
                 <ActionPanel.Section>
                   <RunScriptAction scriptPath={script.filePath} />
+                  <Action.CreateQuicklink
+                    quicklink={{
+                      name: script.name,
+                      link: `raycast://extensions/pomdtr/script-kit/run-script?context=${encodeURIComponent(
+                        JSON.stringify({
+                          scriptPath: script.filePath,
+                        }),
+                      )}`,
+                    }}
+                  />
                 </ActionPanel.Section>
                 <ActionPanel.Section>
                   <NewScriptAction />
@@ -99,17 +127,16 @@ function EditScriptAction(props: { scriptPath: string }) {
 
 function RunScriptAction(props: { scriptPath: string; icon?: Icon; title?: string; args?: string[] }) {
   return (
-    <ActionPanel.Item
+    <Action
       title={props.title || "Run Script"}
       icon={props.icon || Icon.Terminal}
       onAction={async () => {
-        const kar_path = resolve(kit_path, "kar");
         try {
-          await execa(kar_path, [props.scriptPath, ...(props.args || [])]);
+          await execa("/bin/bash", [kar_path, props.scriptPath, ...(props.args || [])]);
           await closeMainWindow();
           await popToRoot();
         } catch (error) {
-          showToast(ToastStyle.Failure, "Could not connect to Kit", "Please check that Kit is running");
+          showToast(Toast.Style.Failure, "Could not connect to Kit", "Please check that Kit is running");
         }
       }}
     />